+++ src/js/js-partials/openModalFilm.js
@@ -1,6 +1,7 @@
 import * as basicLightbox from "basiclightbox";
 import ApiService from "./API";
 import settings from "./settings";
+import { addWatched, addQueue } from "./onButtonClick";
 const { IMG_URL } = settings;
 
 const getImgPath = imgPath => (!imgPath ? `${noImg}` : `${IMG_URL}${imgPath}`);
@@ -123,54 +124,7 @@ async function onClickCard(e) {
     modal.show();
   }
 
-   addWatched();
-  
-  function addWatched() {
-    const addWatchedBtn = document.querySelector('.btn-watch');
-    addWatchedBtn.addEventListener('click', onAddWatchedClick);
-  }
-  
-  function onAddWatchedClick(e) {
-    if (e.target.nodeName !== "BUTTON") {
-      return
-    }
- 
-    const currentActiveBtn = document.querySelector('.modal__button-active');
-    const addWatchedId = e.currentTarget.parentNode.dataset.id;
-    
-    e.currentTarget.textContent = 'Remove from watched';
-    e.currentTarget.classList.toggle('modal__button-active');
-
-    if (currentActiveBtn) {
-      e.currentTarget.textContent = 'add to Watched';
-    }
-
-    localStorage.setItem('watchedMovies', JSON.stringify(addWatchedId)); 
-  }
+  addWatched();
 
   addQueue();
-
-  function addQueue() {
-    const addWatchedBtn = document.querySelector('.btn-queue');
-    addWatchedBtn.addEventListener('click', onAddQueueClick)
-  }
-
-  function onAddQueueClick(e) {
-    if (e.target.nodeName !== "BUTTON") {
-      return
-    }
-
-    const currentActiveBtn = document.querySelector('.modal__button-active');
-    const addQueueId = e.currentTarget.parentNode.dataset.id;
-
-    e.currentTarget.textContent = 'Remove from Queue';
-    e.currentTarget.classList.toggle('modal__button-active');
-
-    if (currentActiveBtn) {
-      e.currentTarget.textContent = 'add to Queue';
-    }
-
-    localStorage.setItem('queueMovies', JSON.stringify(addQueueId)); 
-  }
-  
 }
